Main refactor
