+++ src/crafting/UI/LogicGroupComboBox.java
@@ -1,6 +1,5 @@
 package crafting.UI;
 
-import crafting.Main;
 import java.awt.Color;
 import java.awt.Component;
 import java.awt.Cursor;
