+++ src/crafting/UI/ModifierComboBox.java
@@ -1,7 +1,6 @@
 package crafting.UI;
 
 import crafting.Filters;
-import crafting.Main;
 import java.awt.Color;
 import java.awt.Component;
 import java.awt.event.FocusEvent;
