+++ src/crafting/UI/NumFieldKeyListener.java
@@ -5,7 +5,6 @@
  */
 package crafting.UI;
 
-import crafting.Main;
 import java.awt.event.KeyEvent;
 import java.awt.event.KeyListener;
